requirements.txt
# Core Dependencies
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0

# LLM and Agent Framework
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
agentops>=0.2.0

# File Processing
openpyxl>=3.0.9
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=9.0.0

# Security and Encryption  
cryptography>=3.4.8

# Data Validation
pydantic>=1.10.0

# Templating
jinja2>=3.1.0

# UI and Visualization
streamlit>=1.25.0
plotly>=5.15.0
tabulate>=0.9.0

# Utilities
uuid
pathlib
zipfile
tempfile
hashlib
datetime
logging
dataclasses
typing
re
io
subprocess
sys
os
json